tests/owners: Move ownership change tests to dedicated modules
