+++ src/tests/owners.rs
@@ -8,14 +8,12 @@ use crate::{
 use crates_io::{
     models::Crate,
     views::{
-        EncodableCrateOwnerInvitation, EncodableCrateOwnerInvitationV1, EncodableOwner,
-        EncodablePublicUser, InvitationResponse,
+        EncodableCrateOwnerInvitationV1, EncodableOwner, EncodablePublicUser, InvitationResponse,
     },
     Emails,
 };
 
 use chrono::{Duration, Utc};
-use crates_io::models::token::{CrateScope, EndpointScope};
 use diesel::prelude::*;
 use http::StatusCode;
 use insta::assert_display_snapshot;
@@ -254,288 +252,6 @@ fn modify_multiple_owners() {
     assert_eq!(app.db(|conn| krate.owners(conn).unwrap()).len(), 3);
 }
 
-#[test]
-fn owner_change_via_cookie() {
-    let (app, _, cookie) = TestApp::full().with_user();
-
-    let user2 = app.db_new_user("user-2");
-    let user2 = user2.as_model();
-
-    let krate =
-        app.db(|conn| CrateBuilder::new("foo_crate", cookie.as_model().id).expect_build(conn));
-
-    let url = format!("/api/v1/crates/{}/owners", krate.name);
-    let body = json!({ "owners": [user2.gh_login] });
-    let body = serde_json::to_vec(&body).unwrap();
-    let response = cookie.put::<()>(&url, body);
-    assert_eq!(response.status(), StatusCode::OK);
-    assert_eq!(
-        response.json(),
-        json!({ "ok": true, "msg": "user user-2 has been invited to be an owner of crate foo_crate" })
-    );
-}
-
-#[test]
-fn owner_change_via_token() {
-    let (app, _, _, token) = TestApp::full().with_token();
-
-    let user2 = app.db_new_user("user-2");
-    let user2 = user2.as_model();
-
-    let krate =
-        app.db(|conn| CrateBuilder::new("foo_crate", token.as_model().user_id).expect_build(conn));
-
-    let url = format!("/api/v1/crates/{}/owners", krate.name);
-    let body = json!({ "owners": [user2.gh_login] });
-    let body = serde_json::to_vec(&body).unwrap();
-    let response = token.put::<()>(&url, body);
-    assert_eq!(response.status(), StatusCode::OK);
-    assert_eq!(
-        response.json(),
-        json!({ "ok": true, "msg": "user user-2 has been invited to be an owner of crate foo_crate" })
-    );
-}
-
-#[test]
-fn owner_change_via_change_owner_token() {
-    let (app, _, _, token) =
-        TestApp::full().with_scoped_token(None, Some(vec![EndpointScope::ChangeOwners]));
-
-    let user2 = app.db_new_user("user-2");
-    let user2 = user2.as_model();
-
-    let krate =
-        app.db(|conn| CrateBuilder::new("foo_crate", token.as_model().user_id).expect_build(conn));
-
-    let url = format!("/api/v1/crates/{}/owners", krate.name);
-    let body = json!({ "owners": [user2.gh_login] });
-    let body = serde_json::to_vec(&body).unwrap();
-    let response = token.put::<()>(&url, body);
-    assert_eq!(response.status(), StatusCode::OK);
-    assert_eq!(
-        response.json(),
-        json!({ "ok": true, "msg": "user user-2 has been invited to be an owner of crate foo_crate" })
-    );
-}
-
-#[test]
-fn owner_change_via_change_owner_token_with_matching_crate_scope() {
-    let crate_scopes = Some(vec![CrateScope::try_from("foo_crate").unwrap()]);
-    let endpoint_scopes = Some(vec![EndpointScope::ChangeOwners]);
-    let (app, _, _, token) = TestApp::full().with_scoped_token(crate_scopes, endpoint_scopes);
-
-    let user2 = app.db_new_user("user-2");
-    let user2 = user2.as_model();
-
-    let krate =
-        app.db(|conn| CrateBuilder::new("foo_crate", token.as_model().user_id).expect_build(conn));
-
-    let url = format!("/api/v1/crates/{}/owners", krate.name);
-    let body = json!({ "owners": [user2.gh_login] });
-    let body = serde_json::to_vec(&body).unwrap();
-    let response = token.put::<()>(&url, body);
-    assert_eq!(response.status(), StatusCode::OK);
-    assert_eq!(
-        response.json(),
-        json!({ "ok": true, "msg": "user user-2 has been invited to be an owner of crate foo_crate" })
-    );
-}
-
-#[test]
-fn owner_change_via_change_owner_token_with_wrong_crate_scope() {
-    let crate_scopes = Some(vec![CrateScope::try_from("bar").unwrap()]);
-    let endpoint_scopes = Some(vec![EndpointScope::ChangeOwners]);
-    let (app, _, _, token) = TestApp::full().with_scoped_token(crate_scopes, endpoint_scopes);
-
-    let user2 = app.db_new_user("user-2");
-    let user2 = user2.as_model();
-
-    let krate =
-        app.db(|conn| CrateBuilder::new("foo_crate", token.as_model().user_id).expect_build(conn));
-
-    let url = format!("/api/v1/crates/{}/owners", krate.name);
-    let body = json!({ "owners": [user2.gh_login] });
-    let body = serde_json::to_vec(&body).unwrap();
-    let response = token.put::<()>(&url, body);
-    assert_eq!(response.status(), StatusCode::FORBIDDEN);
-    assert_eq!(
-        response.json(),
-        json!({ "errors": [{ "detail": "must be logged in to perform that action" }] })
-    );
-}
-
-#[test]
-fn owner_change_via_publish_token() {
-    let (app, _, _, token) =
-        TestApp::full().with_scoped_token(None, Some(vec![EndpointScope::PublishUpdate]));
-
-    let user2 = app.db_new_user("user-2");
-    let user2 = user2.as_model();
-
-    let krate =
-        app.db(|conn| CrateBuilder::new("foo_crate", token.as_model().user_id).expect_build(conn));
-
-    let url = format!("/api/v1/crates/{}/owners", krate.name);
-    let body = json!({ "owners": [user2.gh_login] });
-    let body = serde_json::to_vec(&body).unwrap();
-    let response = token.put::<()>(&url, body);
-    assert_eq!(response.status(), StatusCode::FORBIDDEN);
-    assert_eq!(
-        response.json(),
-        json!({ "errors": [{ "detail": "must be logged in to perform that action" }] })
-    );
-}
-
-#[test]
-fn owner_change_without_auth() {
-    let (app, anon, cookie) = TestApp::full().with_user();
-
-    let user2 = app.db_new_user("user-2");
-    let user2 = user2.as_model();
-
-    let krate =
-        app.db(|conn| CrateBuilder::new("foo_crate", cookie.as_model().id).expect_build(conn));
-
-    let url = format!("/api/v1/crates/{}/owners", krate.name);
-    let body = json!({ "owners": [user2.gh_login] });
-    let body = serde_json::to_vec(&body).unwrap();
-    let response = anon.put::<()>(&url, body);
-    assert_eq!(response.status(), StatusCode::FORBIDDEN);
-    assert_eq!(
-        response.json(),
-        json!({ "errors": [{ "detail": "must be logged in to perform that action" }] })
-    );
-}
-
-#[test]
-fn test_owner_change_with_legacy_field() {
-    let (app, _, user1) = TestApp::full().with_user();
-    app.db(|conn| CrateBuilder::new("foo", user1.as_model().id).expect_build(conn));
-    app.db_new_user("user2");
-
-    let input = r#"{"users": ["user2"]}"#;
-    let response = user1.put::<()>("/api/v1/crates/foo/owners", input.as_bytes());
-    assert_eq!(response.status(), StatusCode::OK);
-    assert_display_snapshot!(response.text(), @r###"{"msg":"user user2 has been invited to be an owner of crate foo","ok":true}"###);
-}
-
-#[test]
-fn test_owner_change_with_invalid_json() {
-    let (app, _, user) = TestApp::full().with_user();
-    app.db_new_user("bar");
-    app.db(|conn| CrateBuilder::new("foo", user.as_model().id).expect_build(conn));
-
-    // incomplete input
-    let input = r#"{"owners": ["foo", }"#;
-    let response = user.put::<()>("/api/v1/crates/foo/owners", input.as_bytes());
-    assert_eq!(response.status(), StatusCode::BAD_REQUEST);
-    assert_display_snapshot!(response.text(), @r###"{"errors":[{"detail":"Failed to parse the request body as JSON: owners[1]: expected value at line 1 column 20"}]}"###);
-
-    let response = user.delete_with_body::<()>("/api/v1/crates/foo/owners", input.as_bytes());
-    assert_eq!(response.status(), StatusCode::BAD_REQUEST);
-    assert_display_snapshot!(response.text(), @r###"{"errors":[{"detail":"Failed to parse the request body as JSON: owners[1]: expected value at line 1 column 20"}]}"###);
-
-    // `owners` is not an array
-    let input = r#"{"owners": "foo"}"#;
-    let response = user.put::<()>("/api/v1/crates/foo/owners", input.as_bytes());
-    assert_eq!(response.status(), StatusCode::UNPROCESSABLE_ENTITY);
-    assert_display_snapshot!(response.text(), @r###"{"errors":[{"detail":"Failed to deserialize the JSON body into the target type: owners: invalid type: string \"foo\", expected a sequence at line 1 column 16"}]}"###);
-
-    let response = user.delete_with_body::<()>("/api/v1/crates/foo/owners", input.as_bytes());
-    assert_eq!(response.status(), StatusCode::UNPROCESSABLE_ENTITY);
-    assert_display_snapshot!(response.text(), @r###"{"errors":[{"detail":"Failed to deserialize the JSON body into the target type: owners: invalid type: string \"foo\", expected a sequence at line 1 column 16"}]}"###);
-
-    // missing `owners` and/or `users` fields
-    let input = r#"{}"#;
-    let response = user.put::<()>("/api/v1/crates/foo/owners", input.as_bytes());
-    assert_eq!(response.status(), StatusCode::UNPROCESSABLE_ENTITY);
-    assert_display_snapshot!(response.text(), @r###"{"errors":[{"detail":"Failed to deserialize the JSON body into the target type: missing field `owners` at line 1 column 2"}]}"###);
-
-    let response = user.delete_with_body::<()>("/api/v1/crates/foo/owners", input.as_bytes());
-    assert_eq!(response.status(), StatusCode::UNPROCESSABLE_ENTITY);
-    assert_display_snapshot!(response.text(), @r###"{"errors":[{"detail":"Failed to deserialize the JSON body into the target type: missing field `owners` at line 1 column 2"}]}"###);
-}
-
-#[test]
-fn invite_already_invited_user() {
-    let (app, _, _, owner) = TestApp::init().with_token();
-    app.db_new_user("invited_user");
-    app.db(|conn| CrateBuilder::new("crate_name", owner.as_model().user_id).expect_build(conn));
-
-    // Ensure no emails were sent up to this point
-    assert_eq!(app.as_inner().emails.mails_in_memory().unwrap().len(), 0);
-
-    // Invite the user the first time
-    let response = owner.add_named_owner("crate_name", "invited_user");
-    assert_eq!(response.status(), StatusCode::OK);
-    assert_eq!(
-        response.json(),
-        json!({
-            "msg": "user invited_user has been invited to be an owner of crate crate_name",
-            "ok": true,
-        })
-    );
-
-    // Check one email was sent, this will be the ownership invite email
-    assert_eq!(app.as_inner().emails.mails_in_memory().unwrap().len(), 1);
-
-    // Then invite the user a second time, the message should point out the user is already invited
-    let response = owner.add_named_owner("crate_name", "invited_user");
-    assert_eq!(response.status(), StatusCode::OK);
-    assert_eq!(
-        response.json(),
-        json!({
-            "msg": "user invited_user already has a pending invitation to be an owner of crate crate_name",
-            "ok": true,
-        })
-    );
-
-    // Check that no new email is sent after the second invitation
-    assert_eq!(app.as_inner().emails.mails_in_memory().unwrap().len(), 1);
-}
-
-#[test]
-fn invite_with_existing_expired_invite() {
-    let (app, _, _, owner) = TestApp::init().with_token();
-    app.db_new_user("invited_user");
-    let krate =
-        app.db(|conn| CrateBuilder::new("crate_name", owner.as_model().user_id).expect_build(conn));
-
-    // Ensure no emails were sent up to this point
-    assert_eq!(app.as_inner().emails.mails_in_memory().unwrap().len(), 0);
-
-    // Invite the user the first time
-    let response = owner.add_named_owner("crate_name", "invited_user");
-    assert_eq!(response.status(), StatusCode::OK);
-    assert_eq!(
-        response.json(),
-        json!({
-            "msg": "user invited_user has been invited to be an owner of crate crate_name",
-            "ok": true,
-        })
-    );
-
-    // Check one email was sent, this will be the ownership invite email
-    assert_eq!(app.as_inner().emails.mails_in_memory().unwrap().len(), 1);
-
-    // Simulate the previous invite expiring
-    expire_invitation(&app, krate.id);
-
-    // Then invite the user a second time, a new invite is created as the old one expired
-    let response = owner.add_named_owner("crate_name", "invited_user");
-    assert_eq!(response.status(), StatusCode::OK);
-    assert_eq!(
-        response.json(),
-        json!({
-            "msg": "user invited_user has been invited to be an owner of crate crate_name",
-            "ok": true,
-        })
-    );
-
-    // Check that the email for the second invite was sent
-    assert_eq!(app.as_inner().emails.mails_in_memory().unwrap().len(), 2);
-}
-
 /// Testing the crate ownership between two crates and one team.
 /// Given two crates, one crate owned by both a team and a user,
 /// one only owned by a user, check that the CrateList returned
@@ -627,17 +343,6 @@ fn test_unknown_crate() {
     assert_display_snapshot!(response.text(), @r###"{"errors":[{"detail":"crate `unknown` does not exist"}]}"###);
 
     let response = user.get::<()>("/api/v1/crates/unknown/owner_user");
-    assert_eq!(response.status(), StatusCode::NOT_FOUND);
-    assert_display_snapshot!(response.text(), @r###"{"errors":[{"detail":"crate `unknown` does not exist"}]}"###);
-
-    let body = json!({ "owners": ["bar"] });
-    let body = serde_json::to_vec(&body).unwrap();
-
-    let response = user.put::<()>("/api/v1/crates/unknown/owners", body.clone());
-    assert_eq!(response.status(), StatusCode::NOT_FOUND);
-    assert_display_snapshot!(response.text(), @r###"{"errors":[{"detail":"crate `unknown` does not exist"}]}"###);
-
-    let response = user.delete_with_body::<()>("/api/v1/crates/unknown/owners", body);
     assert_eq!(response.status(), StatusCode::NOT_FOUND);
     assert_display_snapshot!(response.text(), @r###"{"errors":[{"detail":"crate `unknown` does not exist"}]}"###);
 }
