+++ src/tests/routes/crates/owners/mod.rs
@@ -1,1 +1,2 @@
 mod add;
+mod remove;
